Changes related to Sonar Cube code analysis
